### 2022-11-25 #4965 @editorialbot
[PRE REVIEW]: Pybullet Industrial: A process-aware robot simulation
**Submitting author:** @liquidcronos (<a href="http://orcid.org/0000-0002-7825-3476">Jan Baumgärtner</a>)
**Repository:** https://github.com/WBK-Robotics/pybullet_industrial
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
