### 2022-11-25 #4965 @arfon
@liquidcronos – could you merge your default git branch into the `paper` branch so our bot can pick up the code statistics here please?
