### 2022-11-27 #4965 @arfon
Output of `cloc` for @gkthiruvathukal:

```
cloc pybullet_industrial/
      87 text files.
      87 unique files.
       4 files ignored.

Language                     files          blank        comment           code
Python                          37            784            756           2600
XML                             28             93              5           2391
reStructuredText                 5            203            153            287
SVG                              2              0              2            177
Markdown                         6             52              0            138
YAML                             4             28             44            120
DOS Batch                        1              8              1             26
make                             1              4              7              9
SUM:                            84           1172            968           5748
```
