### 2022-12-26 #4965 @arfon
:wave @adi3 – this looks like it might be up your street?
