### 2023-01-10 e8f6574
initialwork
This is the work for the Gcode_class which as been done before joining the Github project
