+++ examples/code.txt
@@ -1,0 +1,1 @@
+G1 X1.9 Y-0.5 Z1.5 

[…]
