### 2023-02-15 0940da0
added code for minimum bug example
