### 2023-02-15 1cae86e
tested how bug behaves with different endeffectors
