+++ minimum_bug_example.py
@@ -9,7 +9,7 @@ if __name__ == "__main__":
     file_directory = os.path.dirname(os.path.abspath(__file__))
     sdmbot_urdf_file = os.path.join(file_directory, 'urdf', 'sdmbot.urdf')
     endeffector_file = os.path.join(
-        file_directory, 'urdf', 'welding_torch.urdf')
+        file_directory,'examples', 'robot_descriptions', 'milling_head.urdf')
 
     physics_client = p.connect(p.GUI)
     p.configureDebugVisualizer(p.COV_ENABLE_GUI, 0)
@@ -26,9 +26,15 @@ if __name__ == "__main__":
     test_tool.set_tool_pose([0, 0, 0.63], [0, 0, 0, 1])
     endeffector.couple(robot)
 
-    position = np.array([0, 0, 0.63])
+    desired_position = np.array([0, 0, 0.63])
 
+    for _ in range(100):
+        endeffector.set_tool_pose(desired_position, [0, 0, 0, 1])
+        p.stepSimulation()
+
+    actual_position,_=endeffector.get_tool_pose()
+    print(actual_position,desired_position)
     while True:
-        endeffector.set_tool_pose(position, [0, 0, 0, 1])
+        endeffector.set_tool_pose(desired_position, [0, 0, 0, 1])
         p.stepSimulation()
         time.sleep(0.01)
