### 2023-02-15 f76c68a
tested joint configurations with welding torch
