### 2023-02-19 #5174 @arfon
@CameronDevine, @sea-bass This is the review thread for the paper. All of the communications will happen here from now on.

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention `https://github.com/openjournals/joss-reviews/issues/5174` so that a link is created to this thread (and @adi3 can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.
