### 2023-03-05 WBK-Robotics/pybullet_industrial#40 @sea-bass
[JOSS REVIEW] Paper feedback
Review thread at : https://github.com/openjournals/joss-reviews/issues/5174

Great job submitting Pybullet_industrial to JOSS -- I had fun trying it out. This issue captures my feedback on your paper submission.

# High-level comments
## Overall
* The paper title stylizes your tool as "Pybullet Industrial", but both the body of the text and this repo does "Pybullet_industrial". Furthermore, PyBullet is officially stylized with a capital B. So I think all the resources should reflect that and consistently use "PyBullet_industrial".
* The citation count seems a bit low, but I have some suggestions throughout this review.
* I think after the "References" section, you could stand to add more content showing actual examples of your tool in action. It's great that you have listed all the features of Pybullet_industrial, but most papers have something akin to an "Experiments" section. Make sure you include screenshots of some of your key examples, whether they are the ones shipping in the repo itself, or maybe some of the ways you have used the tool for your actual research. Bonus points if you link to some animations/videos as well.
* Finally, you should also minimally include a Conclusion section that summarizes your work and its significance, and outlines some directions of future work you are considering and/or actively working on.

## Summary
* It would be helpful to provide some more detail here, even if a sentence or two, describing what is a "process simulation". Do you mean manufacturing processes? Since the reference count is low, maybe this is an opportunity to throw in a reference here as well.
* Add a citation for G-code interpolation. Maybe consider something like [this reference](https://www.nist.gov/publications/nist-rs274ngc-interpreter-version-3) and include the term RS274 which is the official standard name? You may also want to explicitly call out your supported linear, arc, and spline capabilities.

## Overview
* In "Robot Objects" section, could you expand more on what the convenient interfaces involve? What are you using for forward/inverse kinematics, and is there room for citations of those methods? You could also cite URDF itself with e.g. [this](https://cs.paperswithcode.com/paper/understanding-urdf-a-survey-based-on-user).
* In "Endeffector Tools", consider citing ray casting with [the original paper](https://www.sciencedirect.com/science/article/abs/pii/0146664X82901691?via%3Dihub)
* You should absolutely talk about the Extruder and Finger Gripper objects in the paper as well! I found them later in the docs and they are a key part of your work!

---

# Grammar/typos
## General
* Wherever you mention Python, make sure it is upper-case P
* Wherever you mention PyBullet, make sure it is upper-case P and B
* You should replace "endeffector" with "end effector" throughout the paper

## Statement of Need
* Missing space between webots and the citation
*  "Pybullet_industrial is thus the first process-aware robot simulation platform build for research." --> "built for research"
* In the bulleted list, you mix conjugation "Designing, Design, Check". Either do "Design, Design, Check" or "Designing, Designing, Checking"
