### 2023-03-05 #5174 @sea-bass
Hello all, my review is complete.

Overall, the software is in very good shape and checks all the boxes. Only found some minor things, but getting set up with the tool was easy and the claims were verifiable.

The paper, on the other hand, is a bit short on references, but critically it abruptly stops after outlining many of their features (they could do a bit more on that front). So I think my biggest piece of feedback to the authors is to add a section showing examples/applications of their work in action, plus a conclusion summarizing their work and outlining future work.

See https://github.com/WBK-Robotics/pybullet_industrial/issues/40 for documentation review and https://github.com/WBK-Robotics/pybullet_industrial/issues/41 for software review. Once my comments are addressed, I'd be happy to take another quick look and help move this forward!
