### 2023-03-07 038fb2c
added Example section to paper
