### 2023-03-07 4c323fb
added conclusion, more references and overview of robot interfaces
