+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Pybullet Industrial: A process-aware robot simulation'
+title: 'PyBullet Industrial: A process-aware robot simulation'
 tags:
   - Python
   - robotics
@@ -32,25 +32,25 @@ bibliography: paper.bib
 The trend towards individualized products and the increasing demand for a greater number of variants require a rethinking in the production engineering environment. In the context of this transformation, we see robots taking on more and more manufacturing tasks [@wsk].
 The development of this field is hampered by a toolchain gap: While there are a large number of robot simulations and process simulations there is not yet a simple simulation environment that combines the two and allows the user to investigate the interplay of both.
 
-To meet this challenge we developed Pybullet_industrial. This python package extends the open source multi body physics package Pybullet with manufacturing process models to simulate manufacturing applications that add material, remove material or simply move material.
+To meet this challenge we developed PyBullet Industrial. This Python package extends the open source multi body physics package PyBullet with manufacturing process models to simulate manufacturing applications that add material, remove material or simply move material.
 A sample of concrete manufacturing applications in each category can be seen in Figure \ref{manu_process}.
 
-The package not only simulates the environmental effect of the processes but also the forces imparted on the robot. It also allows the dynamic switching of processes with the same robot corresponding to tool changes during the manufacturing process. The package also contains utility functions such as path builder classes which are based on G-code interpolation schemes or a variety of drawing and visualization functions.
+The package not only simulates the environmental effect of the processes but also the forces imparted on the robot. It also allows the dynamic switching of processes with the same robot corresponding to tool changes during the manufacturing process. The package also contains utility functions such as path builder classes which are based on G-code (also called RS274) interpolation schemes [@g-code] or a variety of drawing and visualization functions.
 
 
 
 # Statement of need
-Pybullet_industrial was developed for the interdisciplinary field of robot manufacturing.
+PyBullet Industrial was developed for the interdisciplinary field of robot manufacturing.
 While there are a large number of simulation tools for robotics research such as Gazebo [@gazebo], CoppeliaSim [@coppeliasim], or webots[@webots], their capabilities all end at the robot endeffector as they are unable to simulate manufacturing processes.
 In the same vein, there are several popular FE simulation tools such as Abaqus [@abaqus] capable of simulating process behavior.
 These simulations end at the tool as they are not meant to simulate the systems that move the tool.
 Since robots are now performing more and more manufacturing tasks studying and accounting for the interaction between robots and processes becomes ever more important.
 This requires a simulation that can simulate robots and processes.
 
-Pybullet_industrial closes this gap by taking classical robot multibody simulations and extending them using simple process simulations which impact the environment.
-Pybullet_industrial is thus the first process-aware robot simulation platform build for research.
-Note that Pybullet_industrial neither aims to develop perfect process simulations nor robot simulations, it focuses on the interplay of both.
-Example applications of Pybullet_industrial are:
+PyBullet Industrial closes this gap by taking classical robot multibody simulations and extending them using simple process simulations which impact the environment.
+PyBullet Industrial is thus the first process-aware robot simulation platform build for research.
+Note that PyBullet Industrial neither aims to develop perfect process simulations nor robot simulations, it focuses on the interplay of both.
+Example applications of PyBullet Industrial are:
 
 * Designing joint controllers that compensate for the large process forces during milling
 * Design path planning algorithms for 3D printing that can detect if a robot combines with a previously printed object.
@@ -59,16 +59,23 @@ Example applications of Pybullet_industrial are:
 
 # Overview
 
[…]
