+++ paper.bib
@@ -7,6 +7,30 @@
   organization={IEEE}
 }
 
+@misc{g-code,
+  author = {Thomas Kramer and Frederick Proctor and Elena Messina},
+  title = {The NIST RS274NGC Interpreter - Version 3},
+  year = {2000},
+  month = {2000-08-01},
+  publisher = {NIST Interagency/Internal Report (NISTIR), National Institute of Standards and Technology, Gaithersburg, MD},
+  url = {https://tsapps.nist.gov/publication/get_pdf.cfm?pub_id=823374},
+  language = {en},
+}
+
+@article{raycast,
+title = {Ray casting for modeling solids},
+journal = {Computer Graphics and Image Processing},
+volume = {18},
+number = {2},
+pages = {109-144},
+year = {1982},
+issn = {0146-664X},
+doi = {https://doi.org/10.1016/0146-664X(82)90169-1},
+url = {https://www.sciencedirect.com/science/article/pii/0146664X82901691},
+author = {Scott D Roth},
+abstract = {This paper presents ray casting as the methodological basis for a CAD/CAM solid modeling system. Solid objects are modeled by combining primitive solids, such as blocks and cylinders, using the set operators union, intersection, and difference. To visualize and analyze the composite solids modeled, virtual light rays are cast as probes. By virtue of its simplicity, ray casting is reliable and extensible. The most difficult mathematical problem is finding line-surface intersection points. So surfaces such as planes, quadrics, tori, and probably even parametric surface patches may bound the primitive solids. The adequacy and efficiency of ray casting are issues addressed here. A fast picture generation capability for interactive modeling is the biggest challenge. New methods are presented, accompanied by sample pictures and CPU times, to meet the challenge.}
+}
+
 \bibitem{wsk}
 Mühlbeier, E., Gönnheimer, P., Hausmann, L., Fleischer, J.: Value Stream Kinematics. Production at the leading edge of technology. WGP 2020. Lecture Notes in Production Engineering. \url{doi: 10.1007/978-3-662-62138-7_41}
 
