### 2023-03-07 cda5933
adressed concerns in #40
