+++ paper.md
@@ -31,8 +31,11 @@ bibliography: paper.bib
 # Summary
 The trend towards individualized products and the increasing demand for a greater number of variants require a rethinking in the production engineering environment. In the context of this transformation, we see robots taking on more and more manufacturing tasks [@wsk].
 The development of this field is hampered by a toolchain gap: While there are a large number of robot simulations and process simulations there is not yet a simple simulation environment that combines the two and allows the user to investigate the interplay of both.
+Process simulation in this case refers to the simulation of manufacturing processes which according to [@process_sim] is defined as the use of one or
+more physical mechanisms to transform the shape and/or form of a workpiece.
+While process simulation is a vast field that studies different levels of detail, we focus on the simulation of how the process affects the environment and the robot.
 
-To meet this challenge we developed PyBullet Industrial. This Python package extends the open source multi body physics package PyBullet with manufacturing process models to simulate manufacturing applications that add material, remove material or simply move material.
+To meet this challenge we developed PyBullet Industrial. This Python package extends the open-source multi-body physics package PyBullet with manufacturing process models to simulate manufacturing applications that add material, remove material or simply move material.
 A sample of concrete manufacturing applications in each category can be seen in Figure \ref{manu_process}.
 
 The package not only simulates the environmental effect of the processes but also the forces imparted on the robot. It also allows the dynamic switching of processes with the same robot corresponding to tool changes during the manufacturing process. The package also contains utility functions such as path builder classes which are based on G-code (also called RS274) interpolation schemes [@g-code] or a variety of drawing and visualization functions.
@@ -41,25 +44,25 @@ The package not only simulates the environmental effect of the processes but als
 
 # Statement of need
 PyBullet Industrial was developed for the interdisciplinary field of robot manufacturing.
-While there are a large number of simulation tools for robotics research such as Gazebo [@gazebo], CoppeliaSim [@coppeliasim], or webots[@webots], their capabilities all end at the robot endeffector as they are unable to simulate manufacturing processes.
+While there are a large number of simulation tools for robotics research such as Gazebo [@gazebo], CoppeliaSim [@coppeliasim], or webots [@webots], their capabilities all end at the robot end effector as they are unable to simulate manufacturing processes.
 In the same vein, there are several popular FE simulation tools such as Abaqus [@abaqus] capable of simulating process behavior.
 These simulations end at the tool as they are not meant to simulate the systems that move the tool.
 Since robots are now performing more and more manufacturing tasks studying and accounting for the interaction between robots and processes becomes ever more important.
 This requires a simulation that can simulate robots and processes.
 
 PyBullet Industrial closes this gap by taking classical robot multibody simulations and extending them using simple process simulations which impact the environment.
-PyBullet Industrial is thus the first process-aware robot simulation platform build for research.
+PyBullet Industrial is thus the first process-aware robot simulation platform built for research.
 Note that PyBullet Industrial neither aims to develop perfect process simulations nor robot simulations, it focuses on the interplay of both.
 Example applications of PyBullet Industrial are:
 
 * Designing joint controllers that compensate for the large process forces during milling
-* Design path planning algorithms for 3D printing that can detect if a robot combines with a previously printed object.
-* Check the coating of an object in complex scenarios where the object is moved by a robot while another one is spraying paint.
+* Designing path planning algorithms for 3D printing that can detect if a robot combines with a previously printed object.
+* Checking the coating of an object in complex scenarios where the object is moved by a robot while another one is spraying paint.
 
 
 # Overview
 
[…]
