+++ paper.bib
@@ -17,6 +17,31 @@
   language = {en},
 }
 
+@article{process_sim,
+title = {Simulation in Manufacturing: Review and Challenges},
+journal = {Procedia CIRP},
+volume = {25},
+pages = {213-229},
+year = {2014},
+note = {8th International Conference on Digital Enterprise Technology - DET 2014 Disruptive Innovation in Manufacturing Engineering towards the 4th Industrial Revolution},
+issn = {2212-8271},
+doi = {https://doi.org/10.1016/j.procir.2014.10.032},
+url = {https://www.sciencedirect.com/science/article/pii/S2212827114010634},
+author = {D. Mourtzis and M. Doukas and D. Bernidaki},
+keywords = {Manufacturing, Simulation, Information and Communication Technologies},
+abstract = {Simulation comprises an indispensable set of technological tools and methods for the successful implementation of digital manufacturing, since it allows for the experimentation and validation of product, process and system design and configuration. Especially in todays’ turbulent manufacturing environment, which is affected by megatrends such as globalisation and ever-increasing requirements for higher degree of product customisation and personalisation, the value of simulation is evident. This keynote paper investigates the major milestones in the evolution of simulation technologies and examines recent industrial and research applications and findings. Based on this review, the identification of gaps in current practices is presented, and future trends and challenges to be met on the field are outlined. The considered simulation methods and tools include CAx, Factory layout design, Material and Information flow design, Manufacturing Networks Design, Manufacturing Systems Planning and Control, Manufacturing Networks Planning and Control, Augmented and Virtual Reality in product and process design, planning and verification (ergonomics, robotics, etc.). The evolution, advances, current practices and future trends of these technologies, industrial applications and research results are discussed in the context of the contemporary manufacturing industry.}
+}
+
+
+@unknown{urdf,
+author = {Tola, Daniella and Corke, Peter},
+year = {2023},
+month = {02},
+pages = {},
+title = {Understanding URDF: A Survey Based on User Experience},
+doi = {10.48550/arXiv.2302.13442}
+}#
+
 @article{raycast,
 title = {Ray casting for modeling solids},
 journal = {Computer Graphics and Image Processing},
