### 2023-03-07 WBK-Robotics/pybullet_industrial#40 @liquidcronos
Hey @sea-bass, thank you for the thorough review.
I have tried to address these in the latest commit on the paper branch. However, I am a bit confused about the missing Grippers and Extruder since both are mentioned in the paper. Do you mean that I should emphasize them more?
