### 2023-03-07 WBK-Robotics/pybullet_industrial#40 @sea-bass
Thanks for the quick addressing of the reviews @liquidcronos!

> Do you mean that I should emphasize them more?

Yes. You added a citation and figure for `MillingTool`, and it would be really nice to at least have the same for the `Extruder`. The parallel gripper is maybe less important since many other simulators already handle standard rigid-body contacts.
