+++ docs/source/tutorials.rst
@@ -21,10 +21,10 @@ A box with rounded corners can be created by the use of build_box_path().
    :language: python
    :linenos:
    :lines: 1-3, 7-21
-   
+
 For positioning the path an array for the center position is specified along with the dimensions of the box and the radius of its corners.
-The path is rotated by 90 degrees around the x-axis using the transformation function of pybullet. 
-In this case, the path contains 100 points. A circle can be generated similarly. 
+The path is rotated by 90 degrees around the x-axis using the transformation function of pybullet.
+In this case, the path contains 100 points. A circle can be generated similarly.
 After the paths are created, they are drawn. The default color of a path is blue.
 The result looks like this:
 
@@ -50,12 +50,12 @@ Manipulating paths
 Sometimes it can be useful to translate an entire path by a given vector.
 Similarly, you can rotate a path by a prescribed quaternion.
 Finally, a path can be appended or prepended to another path, as shown here:
- 
+
 .. literalinclude:: ../../examples/path_creation.py
    :language: python
    :linenos:
    :lines: 62-69
-   
+
 
 3d printing and milling
 ======================
@@ -74,16 +74,16 @@ By coupling or decoupling a tool to a link from the robot, it can be attached or
    :language: python
    :linenos:
    :lines: 1-43
-   
+
 The next step is defining the path for the tools, as it is done here with a box path.
 
 .. literalinclude:: ../../examples/3d_printing_and_milling.py
    :language: python
    :linenos:
-   :lines: 46-49
-   
-   
-Finally, the tools are translated to follow the path. 
+   :lines: 45-48
+
+
+Finally, the tools are translated to follow the path.
 First, they are positioned at the beginning of the path.
 For each position in the path, the tools are translated and extrude or respectively remove particles.
 
@@ -97,4 +97,4 @@ After they reached the end of the path, the tool will be decoupled and the new t
 .. literalinclude:: ../../examples/3d_printing_and_milling.py
    :language: python
    :linenos:
-   :lines: 59-88
+   :lines: 57-88
