### 2023-03-07 WBK-Robotics/pybullet_industrial#42 @CameronDevine
[JOSS REVIEW] Document Examples
All the examples provided are wonderful; however, there really should be documentation provided to go along with each one. At present, the file names are the only indication of what the examples do. Furthermore, several examples are interactive but there are no instructions about how to interact with them.

openjournals/joss-reviews#5174
