### 2023-03-07 WBK-Robotics/pybullet_industrial#43 @CameronDevine
[JOSS REVIEW] Camera Example Dependancies
The `camera_sensors.py` example needs `matplotlib` to be installed to run. However, it is not a dependency of PyBullet Industrial. I would suggest only saving images if `matplotlib` is installed, or if a command line argument is passed to the example.

This is also mentioned in #41.

https://github.com/openjournals/joss-reviews/issues/5174
