### 2023-03-07 WBK-Robotics/pybullet_industrial#44 @CameronDevine
[JOSS REVIEW] Contributing File
As part of the JOSS review the guide for contributing file should include three things,

> 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

The first two are included in the `CONTRIBUTING.md` file; however, there are no instructions on how to seek support.

https://github.com/openjournals/joss-reviews/issues/5174
