### 2023-03-07 #5174 @CameronDevine
The findings from my review are summarized here:

1. I agree with @sea-bass that the paper should include a conclusion. (See https://github.com/WBK-Robotics/pybullet_industrial/issues/40)
2. The examples are currently undocumented. (See https://github.com/WBK-Robotics/pybullet_industrial/issues/42)
3. One of the examples requires `matplotlib` but it isn't a dependency of the library (See https://github.com/WBK-Robotics/pybullet_industrial/issues/43)
4. The community guidelines do not provide any directions on how to seek support (See https://github.com/WBK-Robotics/pybullet_industrial/issues/44)

Once these issues are resolved I think this will make an excellent contribution to JOSS. This package would have been very useful while completing my PhD thesis.
