+++ CONTRIBUTING.md
@@ -31,6 +31,10 @@ Otherwise the maintainer might suggest changes that should be made to the code b
 These changes might include, deleting unnecessairy comments, deleting unused files or refractoring overly complicated code passages.
 These suggestions are generally designed to ensure the continued usability and maintainability of the code base.
 
+## Seeking support
+If you have questions about how to use this project, please open a GitHub issue using the question template.
+However if something is not working as expected, please open a GitHub issue using the bug report template.
+
 ## Any contributions you make will be under the MIT Software License
 In short, when you submit code changes, your submissions are understood to be under the same [MIT License](http://choosealicense.com/licenses/mit/) that covers the project. Feel free to contact the maintainers if that's a concern.
 
