### 2023-03-08 c12b88f
resolves #44
