+++ CONTRIBUTING.md
@@ -33,7 +33,11 @@ These suggestions are generally designed to ensure the continued usability and m
 
 ## Seeking support
 If you have questions about how to use this project, please open a GitHub issue using the question template.
-However if something is not working as expected, please open a GitHub issue using the bug report template.
+This will help us keep track of how people are using the project, and help us improve the documentation.
+Please dont contact the maintainers directly with questions about how to use the project.
+
+If you dont have a question but something is not working as expected, please open a GitHub issue using the bug report template.
+This will not only help keep track of bugs but only prevents you from wasting time on a bug that is already known.
 
 ## Any contributions you make will be under the MIT Software License
 In short, when you submit code changes, your submissions are understood to be under the same [MIT License](http://choosealicense.com/licenses/mit/) that covers the project. Feel free to contact the maintainers if that's a concern.
