### 2023-03-08 cee984a
resolved #43 and addressed matplotlib issue in #41
