### 2023-03-08 f37defe
adding docstrings to each example explaining their usages. resolves #42
