### 2023-03-08 22ddb65
adjusted code of examples
