+++ .github/workflows/python_package.yml
@@ -16,7 +16,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: [3.6.9]
+        python-version: [3.6.15]
 
     steps:
     - uses: actions/checkout@v2
