### 2023-03-08 #5174 @liquidcronos
Thank you very much for your reviews and the valuable feedback.

I have updated the code to reflect your suggestions, and thus removed the bugs from the examples and added docstrings explaining their usage. You can see this in the Issues you opened. I am currently working on incorporating your feedback into the paper and will share the new version with you shortly
