### 2023-03-08 48791a4
added documentation reference and made source installation clearer for #41
