+++ README.md
@@ -17,6 +17,9 @@ With the help of the pybullet_industrial package you will be able to:
 - simulating paint coating scenarios
 - simulate the handling of complex tasks using a variety of grippers
 
+## Documentation
+
+You can find the documentation [here](https://pybullet-industrial.readthedocs.io/en/latest/).
 
 ## Getting Started
 
@@ -45,6 +48,7 @@ The source version is the most up-to-date version of the library. It is recommen
 ```
 
 Note that if you are on windows, you may need to use the command `pip install src\\` instead.
+And depending on how you have installed your python environment, you may need to first install `wheel` using `pip install wheel`.
 
 
 
