### 2023-03-08 WBK-Robotics/pybullet_industrial#42 @CameronDevine
I added a few comments to commit f37defe with some minor edits.
