### 2023-03-08 f0ccc6c
added image of extruder inner working and clarified geometry
