### 2023-03-08 WBK-Robotics/pybullet_industrial#40 @liquidcronos
Hey @sea-bass , thanks for the clarification.
I have now added a section delving deeper into the working of the end effector.
I have also revised the introduction of the end effector class since this this seems to have caused some confusion.
While the library provides geometry files for possible EndeffectorTools, these are independent of the Extruder, MillingTool, etc...
This means that I can not show an image of the tool.

I hope the new version of the paper makes this more clear.

Let me know if the paper is now to your satisfaction!
