### 2023-03-08 #5174 @liquidcronos
Hey @CameronDevine ,
I have now included a conclusion to the paper along with changes @sea-bass  requested.
Let me know if this resolves your reservations about the paper.
