### 2023-03-08 eeb8c90
added gif of tool to readme and docs resolving #41
