### 2023-03-08 WBK-Robotics/pybullet_industrial#41 @liquidcronos
Hey @sea-bass,
I think I have now included everything you ask for.  Let me know if there is anything else i can do!
