### 2023-03-08 ea78f05
fixed spelling mistakes as documented in #42
