### 2023-03-08 WBK-Robotics/pybullet_industrial#42 @liquidcronos
Thank you for pointing out the spelling mistake!
Regarding the pendulum: the endeffector tool itself can also have joints and links therefore a pendulum shaped endeffector object was loaded and force applied to the tcp corresponding to the tip of the pendulum. I have updated the description to make this more clear. I have also since updated the paper to make it clearer that the endeffector tools can have different shapes and even joints which are loaded from URDF files
