### 2023-03-11 WBK-Robotics/pybullet_industrial#40 @sea-bass
@liquidcronos Thanks for going through these comments. It's looking great!

Just a few more comments.

## Important:
* Figure 1 doesn't show up in the latest proof. Please check this.
* My comment on adding an Examples section wasn't to add a code snippet -- actually, I don't think this is appropriate for this kind of paper. I meant to have a section between your Overview and Conclusion that describes the "real" applications that you have achieved with PyBullet Industrial. Can you describe some of the kinds of examples you have in the repo, and also maybe what you or others have achieved with this tool in your research? If you can, you should include at least one screenshot of the full PyBullet Industrial window running your favorite example, and also cite any publication(s) you or others have that use this software!

## Editorial
* The first bullet in the "Statement of Need" list is missing a full stop (".") at the end.
* First sentence in "Robot Objects": PyBullets URDF -> PyBullet's URDF
* Figure 3 looks really big
* Your "Conclusion:" header is the only header with a semicolon (:) at the end of it. Please remove it.
