### 2023-03-11 WBK-Robotics/pybullet_industrial#41 @sea-bass
@liquidcronos

Thank you! Two comments here on your changes:

* Rather than checking for matplotlib, or warning use they didn't install matplotlib at example running time, why not just add it to your [setup.py](https://github.com/WBK-Robotics/pybullet_industrial/blob/main/src/setup.py#L14)? I would understand if you wanted to keep the installation lightweight, but pybullet is so massive that adding matplotlib won't make a big difference, haha
* Found a small typo in "Example" [here](https://github.com/WBK-Robotics/pybullet_industrial/blob/dc788a474217193f46ea792daade59e4e6d4fb59/examples/robot_frame_visualization.py#L1)
