### 2023-03-12 WBK-Robotics/pybullet_industrial#41 @liquidcronos
Thank you for your suggestions I have adressed these in the newest commit.
Regarding matplotlib you are probably right, but on principle i think we shouldnt install libraries that are not needed for the main functioning of the library since only a small subset of people will use the examples.
