### 2023-03-12 b1d2c31
fixed exmaple section to include real life applications as per #40
