+++ paper.md
@@ -136,7 +136,7 @@ For direct control, a ROS interface will be added to allow the use of ROS contro
 Apart from the example applications mentioned in the statement of need, the library is also already been used in a number of ongoing research projects.
 These include:
 
-* A study on VR based robot programming [@vr_robot_programming] where a welding task was simulated in VR using PyBullet Industrial. The resulting Project Demonstrator can be seen at the Hannover Messe 2023.
+* A study on VR based robot programming where a welding task was simulated in VR using PyBullet Industrial. The resulting Project Demonstrator can be seen at the Hannover Messe 2023.
 * A project on automated Electromotor dissasembly where the simulation is used to validate a given dissasembly plan including for example the milling away of rusted screws.
 
 # References
