### 2023-03-12 WBK-Robotics/pybullet_industrial#40 @liquidcronos
Thank you for the great feedback!

Especially the clarification regarding the example, this makes much more sense.
I have included two applications which are  currently being developed using our library. Unfortunately i can not add citations for these projects as the manuscripts for their respective first publications are not yet finished.  Let me know if there is anything else i can do to improve the paper!
