### 2023-03-12 WBK-Robotics/pybullet_industrial#40 @sea-bass
Thank you!

The Example Research Applications section is a little shorter than I anticipated, but it's the right content I had in mind. I still think it would be nice for this section to have a screenshot showing the **full** PyBullet Industrial GUI window with a robot in an industrial-looking background, etc. vs. the simple trajectory in Figure 5. You just need to open up your favorite example from the actual repo and take a screenshot.

Anyways, the paper is in good shape. My only requirement is that your new section appear before Conclusion, as this should be the last section before references. After that and anything else you'd like to add, I'll approve the review!
