### 2023-03-12 WBK-Robotics/pybullet_industrial#41 @sea-bass
Sounds good to me -- thank you for going through this feedback!
