### 2023-03-12 WBK-Robotics/pybullet_industrial#40 @liquidcronos
Hey @sea-bass,
thanks for the quick reply!
I have now included a screenshot and moved the section infront of the conclusion.
