### 2023-03-12 WBK-Robotics/pybullet_industrial#40 @sea-bass
Your new Figure 1 is great.

Thank you for going through this feedback, and for your work on this software + the JOSS submission. I look forward to seeing this published.
