### 2023-03-13 WBK-Robotics/pybullet_industrial#42 @CameronDevine
Unfortunately, I still don't feel like it is clear what is happening in the `tcp_force.py` example. Looking through the code I think there is a constant force on the pendulum. Could you add another sentence to the documentation describing what forces are applied to the pendulum?
