### 2023-03-13 #5174 @CameronDevine
For me, there is one last outstanding issue.

https://github.com/WBK-Robotics/pybullet_industrial/issues/42#issuecomment-1465760402

Once this is complete I will be more than happy to approve the submission.
