### 2023-03-13 8d9d895
made the force applied to the pendulum more clear
