### 2023-03-13 WBK-Robotics/pybullet_industrial#42 @liquidcronos
Hey @CameronDevine, thanks for the feedback.
I have tried to adresse it in the newest commit.
Let me know if this explains it better.
