### 2023-03-13 WBK-Robotics/pybullet_industrial#42 @CameronDevine
I made a final edit to the `tcp_force.py` example in #46.
