### 2023-03-14 WBK-Robotics/pybullet_industrial#42 @liquidcronos
Yeah, definetly, thank you for the feedback, if you have written it yourself its always great to have someone from outside provide feedback as one always has the disadvantage of knowing what one meant :D
