### 2023-03-15 b779f49
Current State Gcode_class
New:
- new loops in order to ensure correct positioning for the interpolations and G0
- Examples for G0 and G1/2/3
- Gcode_dem that gives a demonstration of how the gcode class can be implemented

TO DO's
- adjusting the style of the code (according PEP8)
- finnishing the test_gcode_class
