+++ examples/gcode_class.py
@@ -2,267 +2,307 @@ import pybullet as p
 import pybullet_industrial as pi
 import numpy as np
 import time
-import copy
-
-
-def create_path(p1,p2,or1):
-    path = pi.linear_interpolation(
-                             np.array(p1), np.array(p2), 10)
-    path.orientations = np.transpose(
-                            [or1] * len(path.orientations[0]))
-    return path
 
 def move_along_path(endeffector: pi.EndeffectorTool, path: pi.ToolPath, stop=False):
-    """Moving a designated endeffector along the provided path.
+    """Move a designated robot along the provided path.
+
     Args:
-        endeffector (pi.EndeffectorTool): Endeffector to be moved.
+        robot (pi.RobotBase): Robot to be moved.
         path (pi.ToolPath): Array of points defining the path.
         stop (bool, optional): Whether or not to stop at the end of the movement.
     """
+    
     for positions, orientations, tool_path in path:
-        endeffector.set_tool_pose(positions, orientations)
-        for _ in range(10):
-            p.stepSimulation()
-    if stop:
-        for _ in range(100):
-            p.stepSimulation()
+        for i in range(10):
+            endeffector.set_tool_pose(positions, orientations)
+            for _ in range(10):
+                p.stepSimulation()
+                time.sleep(0.0001)
+            
+            current_position = endeffector.get_tool_pose()[0]
+            current_orientation = np.array(p.getEulerFromQuaternion(endeffector.get_tool_pose()[1]))
+            
+            position_error = np.linalg.norm(current_position-positions)
+            orientation_error = np.linalg.norm(current_orientation-np.array(p.getEulerFromQuaternion(orientations)))
+
+            if position_error < 0.02 and orientation_error<0.004:
+                break       
 
 def move_robot(robot: pi.RobotBase, path: pi.ToolPath, stop=False):
-    """Moving a designated endeffector along the provided path.
+    """Move a designated robot along the provided path.
+
     Args:
-        endeffector (pi.EndeffectorTool): Endeffector to be moved.
+        robot (pi.RobotBase): Robot to be moved.
         path (pi.ToolPath): Array of points defining the path.
         stop (bool, optional): Whether or not to stop at the end of the movement.
     """
     for positions, orientations, tool_path in path:
-        robot.set_endeffector_pose(positions, orientations)
-        for _ in range(10):
-            p.stepSimulation()
-            time.sleep(0.005)
-    if stop:
-        for _ in range(100):
-            p.stepSimulation()      
+        for i in range(10):
+            robot.set_endeffector_pose(positions, orientations)
+            for _ in range(10):
+                p.stepSimulation()
+                time.sleep(0.0001)
+            
+            current_position = robot.get_endeffector_pose()[0]
+            current_orientation = np.array(p.getEulerFromQuaternion(robot.get_endeffector_pose()[1]))
+            position_error = np.linalg.norm(current_position-positions)
+            orientation_error = np.linalg.norm(current_orientation- np.array(p.getEulerFromQuaternion(orientations)))
+
+            if position_error < 0.02 and orientation_error<0.004:
+                break
+
+def get_active_endeffector(endeffector_list: list):
+    """Get the index of the active endeffector in a list of endeffectors.
+
+    Args:
+        endeffector_list (list): List of pi.EndeffectorTool objects.
+
+    Returns:
+        int: Index of the active endeffector, or -1 if no endeffectors are coupled.
+    """
+    n = 0
+    active_endeffector = -1
+    for i in endeffector_list:
+        if i.is_coupled():
+            active_endeffector = n
+            break
+        n = n + 1
+    return active_endeffector
+
 
 class Gcode_class():
[…]
