+++ examples/test_gcode_class.py
@@ -7,40 +7,27 @@ import pybullet_data
 import pybullet_industrial as pi
 from gcode_class import *
 
-def set_nan(test_array):
-    test_array[:] = np.NaN
-    test_array[0] = ""
-
-def set_g(test_array, val):
-    test_array[1] = val
-
-def set_m(test_array, val):
-    test_array[2] = val
-
-def set_t(test_array, val):
-    test_array[3] = val
-
-def set_xyz(test_array, x,y,z):
-    test_array[4] = x
-    test_array[5] = y
-    test_array[6] = z
-
-
-def set_abc(test_array, a,b,c):
-    test_array[7] = a
-    test_array[8] = b
-    test_array[9] = c
-
-def set_ij(test_array, i, j):
-    test_array[10] = i
-    test_array[11] = j
+def test_final_position(robot: pi.RobotBase, target_position, target_orientation,  pos_precision, ori_precision):
+    within_precision = True
     
+    current_position, current_orientation = robot.get_endeffector_pose()
+    current_orientation = np.array(p.getEulerFromQuaternion(current_orientation))
+    position_error = np.linalg.norm(current_position-target_position)
+    orientation_error = np.linalg.norm(
+            current_orientation-target_orientation)
+    print("Orientation_error: " , orientation_error)
+    within_precision = within_precision and (
+            position_error <= pos_precision) and (orientation_error <= ori_precision)
 
-def set_r(test_array, val):
-    test_array[12] = val
+    return within_precision
 
-def set_f(test_array, val):
-    test_array[13] = val
+def create_command(gcode_object : Gcode_class, commands):
+    text_gcode = "text_gcode.txt"
+    with open(text_gcode, "w") as f:
+        f.writelines(commands)
+    g_code = gcode_object.read_gcode(text_gcode)
+    os.remove(text_gcode)
+    return g_code
 
 class Test_Gcode_class(unittest.TestCase):
     
@@ -48,16 +35,19 @@ class Test_Gcode_class(unittest.TestCase):
         dirname = os.path.dirname(__file__)
         textfile = os.path.join(dirname,'test_code.txt')
         test_objetct = Gcode_class()
-        input_array = test_objetct.read_gcode(textfile)
+        gcode = test_objetct.read_gcode(textfile)
         
-        #Check if Comments work 
-        c = '%Kommentar\n'
-        self.assertEqual(c,input_array[0,0])
-        for i in input_array[0,1:]:
-            self.assertTrue(np.isnan(i)) #Rest of the line has to be NaN
-        #Test if Numbers are stored correctly
-        test_array = np.array(['', 0, 1, 2, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0], dtype=object)
-        self.assertTrue(all(test_array==input_array[1]))
+        # Check that the length of the gcode list is correct
+        self.assertEqual(len(gcode), 3)
+
+        # Check that the first line is correct
+        self.assertEqual(gcode[0], [['G', 0], ['X', 1.0], ['Y', 2.0]])
+
+        # Check that the second line is correct
+        self.assertEqual(gcode[1], [['G', 1], ['X', 2.0], ['Y', 3.0]])
+
+        # Check that the third line is correct
+        self.assertEqual(gcode[2], [['G', 2], ['X', 3.0], ['Y', 4.0], ['Z', 5.0]])
         
         
     def test_run_gcode(self):
@@ -93,33 +83,36 @@ class Test_Gcode_class(unittest.TestCase):
         dirname = os.path.dirname(__file__)
[…]
