+++ examples/gcode_G123.txt
@@ -1,0 +1,58 @@
+% with "%" it is possible to place comments in the G-Code
+
+%Go to default position and adjust the orientation
+G0 X1.9 Y-0.5 Z1.5 A-1.5707 B0 C0
+%Approach Gripper
+G1 X2.7 
+G1 Z1.3
+
+%Toolchange (if programmed)
+T1
+%Move up and adjust orientation 
+G1 Z1.5 A3.1415 B0 C0
+%Go back to the default position
+G1 X1.9 Y-0.5
+
+%Close the gripper
+M10
+%Open the gripper
+M11
+
+%Approach Toolchange position
+G1 X2.7
+G1 Z0.7
+
+
+%Decouple Gripper (if programmed)
+T0
+
+%Move up and adjust orientation
+G1 Z1.9  A-1.5707 B0 C0
+%Go to default position
+G1 X1.9 Y-0.5
+
+
+
+%Zero Offset - Activation
+G54
+%Square movement and orientation flip
+G1 A3.1416
+G1 Z0 A0
+G1 X0.2
+G1 Y0.2
+G1 X0
+G1 Y0
+%Zero Offset - Deactivation
+G500
+
+%Go to default position and adjust the orientation
+G1 X1.9 Y-0.5 Z1.9 A-1.5707 B0 C0
+
+%Circular Interpolation - G2 in X-Y Plane
+G17
+G2 Y0.5 R0.6
+G2 Y-0.5 R0.6
+
+%Circular Interpolation - G3 in X-Y Plane
+G3 Y0.5 R0.6
+G3 Y-0.5 R0.6

[…]
