+++ examples/gcode_dem.py
@@ -1,17 +1,17 @@
-from ctypes.wintypes import POINT
-from time import sleep
-from gcode_class import *
+import time
 import os
 import pybullet as p
 import pybullet_data
 import pybullet_industrial as pi
 import numpy as np
 
-def actuate_gripper(gripper: pi.Gripper,val: int):
+
+def actuate_gripper(gripper: pi.Gripper, val: int):
     gripper.actuate(val)
     for _ in range(25):
         p.stepSimulation()
         time.sleep(0.01)
+
 
 def couple_endeffector(gripper: pi.Gripper, robot: pi.RobotBase, link: chr):
     gripper.couple(robot, 'link6')
@@ -19,16 +19,16 @@ def couple_endeffector(gripper: pi.Gripper, robot: pi.RobotBase, link: chr):
         p.stepSimulation()
         time.sleep(0.01)
 
+
 def decouple_endeffector(gripper: pi.Gripper):
     gripper.decouple()
     for _ in range(25):
         p.stepSimulation()
         time.sleep(0.01)
 
+
 if __name__ == "__main__":
-    
     dirname = os.path.dirname(__file__)
-    
     urdf_file1 = os.path.join(dirname,
                               'robot_descriptions', 'comau_nj290_robot.urdf')
     urdf_file2 = os.path.join(dirname,
@@ -38,18 +38,15 @@ if __name__ == "__main__":
     urdf_file4 = os.path.join(dirname,
                               'robot_descriptions', 'gripper.urdf')
     urdf_files = [urdf_file2, urdf_file3, urdf_file4]
-                              
-    physics_client = p.connect(p.GUI, options='--background_color_red=1 ' +
-                                              '--background_color_green=1 ' +
-                                              '--background_color_blue=1')
-    
+    pysics_client = p.connect(p.GUI, options='--background_color_red=1 ' +
+                                             '--background_color_green=1 ' +
+                                             '--background_color_blue=1')
     p.setPhysicsEngineParameter(numSolverIterations=10000)
     p.configureDebugVisualizer(p.COV_ENABLE_GUI, 0)
     p.setPhysicsEngineParameter(numSolverIterations=10000)
     p.setAdditionalSearchPath(pybullet_data.getDataPath())
     p.setAdditionalSearchPath(pybullet_data.getDataPath())
     p.setGravity(0, 0, -10)
-    
     monastryId = p.createCollisionShape(p.GEOM_MESH,
                                         fileName="samurai_monastry.obj",
                                         flags=p.GEOM_FORCE_CONCAVE_TRIMESH)
@@ -59,37 +56,30 @@ if __name__ == "__main__":
 
     start_orientation = p.getQuaternionFromEuler([0, 0, 0])
     robot = pi.RobotBase(urdf_file1, [0, 0, 0], start_orientation)
-    robot.set_joint_position(({'q2': np.deg2rad(-15.0), 'q3': np.deg2rad(-90.0)}))
-    
+    robot.set_joint_position(({'q2': np.deg2rad(-15.0),
+                               'q3': np.deg2rad(-90.0)}))
     for _ in range(100):
         p.stepSimulation()
 
     start_orientation = p.getQuaternionFromEuler([np.pi, 0, 0])
-    test_gripper = pi.Gripper( urdf_file2, [2.7, -0.5, 1.2], start_orientation)
+    test_gripper = pi.Gripper(urdf_file2, [2.7, -0.5, 1.2], start_orientation)
     endeffector_list = []
     endeffector_list.append(test_gripper)
 
-    #Erstellung von M-Befehlen
+    # Erstellung von M-Befehlen
[…]
