+++ src/pybullet_industrial/__init__.py
@@ -11,3 +11,4 @@ from pybullet_industrial.toolpath import *
 from pybullet_industrial.interpolation import circular_interpolation, linear_interpolation, spline_interpolation
 from pybullet_industrial.path_builders import *
 from pybullet_industrial.milling_tool import *
+from pybullet_industrial.gcode_class import *
