+++ text.txt
@@ -1,0 +1,1 @@
+% This is a commentG0 X1.0 Y2.0G1 X2.0 Y3.0G2 X3.0 Y4.0 Z5.0
