+++ src/pybullet_industrial/gcode_class.py
@@ -4,112 +4,93 @@ import numpy as np
 import time
 
 
-def move_along_path(endeffector: pi.EndeffectorTool,
-                    path: pi.ToolPath, stop=False):
-    """Move a designated robot along the provided path.
-
-    Args:
-        robot (pi.RobotBase): Robot to be moved.
-        path (pi.ToolPath): Array of points defining the path.
-        stop (bool, optional): Whether or not to stop
-        at the end of the movement.
-    """
-    for positions, orientations, tool_path in path:
-
-        for i in range(10):
-            endeffector.set_tool_pose(positions, orientations)
-            for _ in range(10):
-                p.stepSimulation()
-                time.sleep(0.0001)
-            current_position = endeffector.get_tool_pose()[0]
-            current_orientation = np.array(p.getEulerFromQuaternion
-                                           (endeffector.get_tool_pose()[1]))
-            position_error = np.linalg.norm(current_position-positions)
-            orientations_euler = np.array(
-                p.getEulerFromQuaternion(orientations))
-            orientation_error = np.linalg.norm(current_orientation -
-                                               orientations_euler)
-
-            if position_error < 0.02 and orientation_error < 0.004:
-                break
-
-
-def move_robot(robot: pi.RobotBase, path: pi.ToolPath, stop=False):
-    """Move a designated robot along the provided path.
-
-    Args:
-        robot (pi.RobotBase): Robot to be moved.
-        path (pi.ToolPath): Array of points defining the path.
-        stop (bool, optional): Whether or not to stop
-        at the end of the movement.
-    """
-    for positions, orientations, tool_path in path:
-        for i in range(20):
-            robot.set_endeffector_pose(positions, orientations)
-            for _ in range(10):
-                p.stepSimulation()
-                time.sleep(0.0001)
-            current_position = robot.get_endeffector_pose()[0]
-            current_orientation = np.array(p.getEulerFromQuaternion(
-                robot.get_endeffector_pose()[1]))
-            position_error = np.linalg.norm(current_position-positions)
-            orientations_euler = np.array(
-                p.getEulerFromQuaternion(orientations))
-            orientation_error = np.linalg.norm(current_orientation -
-                                               orientations_euler)
-
-            if position_error < 0.02 and orientation_error < 0.004:
-                break
-
-
-def get_active_endeffector(endeffector_list: list):
-    """Get the index of the active endeffector in a list of endeffectors.
-
-    Args:
-        endeffector_list (list): List of pi.EndeffectorTool objects.
-
-    Returns:
-        int: Index of the active endeffector, or -1 if no
-        endeffectors are coupled.
-    """
-    n = 0
-    active_endeffector = -1
-    for i in endeffector_list:
-        if i.is_coupled():
-            active_endeffector = n
-            break
-        n = n + 1
-    return active_endeffector
-
-
 class Gcode_class():
-    def __init__(self, filename=None, robot: pi.RobotBase = None,
-                 endeffector_list=None, m_befehle=None, t_befehle=None):
-        if filename is not None and robot is not None:
-            gcode = self.read_gcode(filename)
-            self.run_gcode(gcode, robot, endeffector_list,
-                           m_befehle, t_befehle)
 
-    def read_gcode(self, filename):
+    def __init__(self, robot: pi.RobotBase = None,
+                 endeffector_list: list = None,
+                 m_commands: list = None,
+                 t_commands: list = None, offset: np.array = None,
+                 plane: int = None, interpolation_steps: int = None,
+                 sleep: int = None):
+        """Initialize a PathMover object with the provided parameters.
+
+        Args:
+            robot (pi.RobotBase): The robot to be moved.
+            filename (str, optional): The name of a G-code file to read.
[…]
