### 2023-03-24 12c7a68
Adjusting Codiga Errors
- Minor Codiga Errors were adjusted
