+++ src/pybullet_industrial/g_code_processor.py
@@ -1,0 +1,460 @@
+import time
+import pybullet as p
+from pybullet_industrial import RobotBase
+from pybullet_industrial import ToolPath
+from pybullet_industrial import linear_interpolation
+from pybullet_industrial import circular_interpolation
+import numpy as np
+
+
+class GCodeProcessor:
+
+    def __init__(self, robot: RobotBase = None,
+                 endeffector_list: list = None,
+                 m_commands: list = None,
+                 t_commands: list = None, offset: np.array = None,
+                 axis: int = None, interpolation_steps: int = None,
+                 sleep: int = None):
+        """Initialize a GCodeProcessor object with the provided parameters.
+
+        Args:
+            robot (pi.RobotBase, optional: The robot which is controlled.
+            filename (str, optional): The name of a G-code file to read.
+            endeffector_list (list, optional): List of end effectors to use.
+            m_commands (list, optional): M-commands to execute.
+            t_commands (list, optional): T-commands to execute.
+            offset (np.array, optional): The offset for the path points.
+            axis (int, optional): The axis around which the circle
+                                  is interpolated.Defaults to 2 which
+                                  corresponds to the z-axis (0=x,1=y)
+            interpolation_steps (int, optional): Number of interpolation steps.
+
+        Returns:
+            None
+        """
+        self.new_point = []
+        self.new_or = []
+        self.last_point = []
+        self.last_or = []
+
+        if m_commands is not None:
+            self.m_commands = m_commands
+
+        if t_commands is not None:
+            self.t_commands = t_commands
+
+        if endeffector_list is not None:
+            self.endeffector_list = endeffector_list
+            self.active_endeffector = self.get_active_endeffector()
+        else:
+            self.endeffector_list = None
+            self.active_endeffector = -1
+
+        if robot is not None:
+            self.robot = robot
+            self.calibrate_tool()
+
+        if offset is not None:
+            self.offset = offset
+        else:
+            self.offset = np.array([[0.0, 0.0, 0.0], [0.0, 0.0, 0.0]])
+
+        if axis is not None:
+            self.axis = axis
+        else:
+            self.axis = 2  # X-Y Axis for the circular interpolation
+
+        if interpolation_steps is not None:
+            self.interpolation_steps = interpolation_steps
+        else:
+            self.interpolation_steps = 10
+
+        if sleep is not None:
+            self.sleep = sleep
+        else:
+            self.sleep = 0.0001
+
+    def read_gcode(self, filename: str):
+        """Reads G-Code row by row and saves the processed Data in
+        a stacked List.
+        Comments that start with % are ignored and all the other data is
+        stored as it gets read in.
+        Every Line in the G-Code resembles the same structure as the text file
+
+        Args:
+            filename (str): Source of information
+
+        Returns:
+            gcode
+        """
+        with open(filename, encoding='utf-8') as f:
+            gcode = []
+
+            # Loop over the lines of the file
+            while True:
+                line = f.readline()
+
+                # If the end of the file is reached, break the loop
+                if not line:
+                    break
+
+                # Initialize a new line as a list
+                new_line = []
+
+                # If the line starts with "%", it is a comment line
+                if line[0] == "%" or len(line) <= 1:
+                    pass
+
+                # Otherwise, split the line into its components and insert
+                # them into the new line
+                else:
+                    # Split the line into its components
+                    data = line.split()
+
+                    # Loop over the components
+                    for i in data:
+                        # Determine the ID of the component
+                        id_val = i[0]
+
+                        # Extract the value of the component
+                        val2 = float(i[1:])
+
+                        if id_val in ["G", "M", "T"]:
[…]
