### 2023-04-19 10cad6a
Deletion of leftover files
- g_code_processor class was duplicated in examples
- test_g_code_processor was duplicated in examples
