+++ examples/test_g_code_processor.py
@@ -1,266 +1,0 @@
-import os
-import unittest
-import time
-import numpy as np
-import pybullet as p
-import pybullet_data
-import pybullet_industrial as pi
-
-
-def couple_endeffector(gripper: pi.Gripper, robot: pi.RobotBase, link: chr):
-
-    gripper.couple(robot, link)
-    for _ in range(25):
-        p.stepSimulation()
-        time.sleep(0.01)
-
-
-def decouple_endeffector(gripper: pi.Gripper):
-
-    gripper.decouple()
-    for _ in range(25):
-        p.stepSimulation()
-        time.sleep(0.01)
-
-
-def check_robot_position(robot: pi.RobotBase, target_position,
-                         target_orientation,  pos_tol, ori_tol):
-
-    within_precision = True
-
-    current_position, current_orientation = robot.get_endeffector_pose()
-    current_orientation = np.array(
-        p.getEulerFromQuaternion(current_orientation))
-
-    position_error = np.linalg.norm(current_position-target_position)
-    orientation_error = np.linalg.norm(current_orientation -
-                                       target_orientation)
-
-    within_precision = within_precision and (
-        position_error <= pos_tol) and (orientation_error <= ori_tol)
-
-    return within_precision
-
-
-def check_tool_position(endeffetor: pi.EndeffectorTool, target_position,
-                        target_orientation,  pos_tol, ori_tol):
-    within_precision = True
-
-    current_position, current_orientation = endeffetor.get_tool_pose()
-    current_orientation = np.array(
-        p.getEulerFromQuaternion(current_orientation))
-
-    position_error = np.linalg.norm(current_position-target_position)
-    orientation_error = np.linalg.norm(
-        current_orientation-target_orientation)
-
-    within_precision = within_precision and (
-        position_error <= pos_tol) and (orientation_error <= ori_tol)
-
-    return within_precision
-
-
-def create_command(gcode_object: pi.GCodeProcessor, commands):
-
-    text_gcode = "text_gcode.txt"
-
-    with open(text_gcode, "w", encoding="utf-8") as f:
-        f.writelines(commands)
-
-    gcode = gcode_object.read_gcode(text_gcode)
-    os.remove(text_gcode)
-
-    return gcode
-
-
-class Test_Gcode_class(unittest.TestCase):
-
-    def test_read_gcode(self):
-
-        # Create test textfile
-        cmd1 = "% This is a comment\n"
-        cmd2 = "G0 X1.0 Y2.0\n"
-        cmd3 = "G1 X2.0 Y3.0\n"
-        cmd4 = "G2 X3.0 Y4.0 Z5.0\n"
-        commands = [cmd1, cmd2, cmd3, cmd4]
-
-        # Creating the G-Code Object
-        gcode_object = pi.GCodeProcessor()
-
-        gcode = create_command(gcode_object, commands)
-
-        # Check that the length of the gcode list is correct
-        self.assertEqual(len(gcode), 3)
-
-        # Check that the first line is correct
-        self.assertEqual(gcode[0], [['G', 0], ['X', 1.0], ['Y', 2.0]])
-
-        # Check that the second line is correct
-        self.assertEqual(gcode[1], [['G', 1], ['X', 2.0], ['Y', 3.0]])
-
-        # Check that the third line is correct
-        self.assertEqual(
-            gcode[2], [['G', 2], ['X', 3.0], ['Y', 4.0], ['Z', 5.0]])
-
-    def test_run_gcode(self):
-
-        dirname = os.path.dirname(__file__)
[…]
