+++ src/pybullet_industrial/g_code_processor.py
@@ -13,9 +13,9 @@ class GCodeProcessor:
                  endeffector_list: list = None,
                  m_commands: list = None,
                  t_commands: list = None, offset: np.array = None,
-                 axis: int = None, interpolation_steps: int = None,
-                 sleep: int = None):
-        """Initialize a GCodeProcessor object with the provided parameters.
+                 axis: int = 2, interpolation_steps: int = 10,
+                 sleep: int = 0.0001):
+        """Initialize a PathMover object with the provided parameters.
 
         Args:
             robot (pi.RobotBase, optional: The robot which is controlled.
@@ -74,9 +74,10 @@ class GCodeProcessor:
         else:
             self.sleep = 0.0001
 
-    def read_gcode(self, filename: str):
+    @staticmethod
+    def read_gcode(filename: str):
         """Reads G-Code row by row and saves the processed Data in
-        a stacked List.
+        a List.
         Comments that start with % are ignored and all the other data is
         stored as it gets read in.
         Every Line in the G-Code resembles the same structure as the text file
@@ -91,8 +92,7 @@ class GCodeProcessor:
             gcode = []
 
             # Loop over the lines of the file
-            while True:
-                line = f.readline()
+            for line in f.readlines():
 
                 # If the end of the file is reached, break the loop
                 if not line:
@@ -101,13 +101,9 @@ class GCodeProcessor:
                 # Initialize a new line as a list
                 new_line = []
 
-                # If the line starts with "%", it is a comment line
-                if line[0] == "%" or len(line) <= 1:
-                    pass
+                # Read in G-Code if line is not a comment and not empty
+                if line[0] != "%" and len(line) > 1:
 
-                # Otherwise, split the line into its components and insert
-                # them into the new line
-                else:
                     # Split the line into its components
                     data = line.split()
 
@@ -144,115 +140,91 @@ class GCodeProcessor:
         # Runs the information out of the input_array
         for cmd in gcode:
 
-            g_com, m_com, t_com = np.nan, np.nan, np.nan
+            cmd_type = cmd[0][0]
 
-            # Checking for the command and advising it to the value
-            if cmd[0][0] == "G":
-                g_com = cmd[0][1]
-            elif cmd[0][0] == "M":
-                m_com = cmd[0][1]
-            elif cmd[0][0] == "T":
-                t_com = cmd[0][1]
+            if cmd_type == "M":
+                self.excecute_m_cmd(cmd)
 
-            x_val, y_val, z_val = np.nan, np.nan, np.nan
-            a_val, b_val, c_val = np.nan, np.nan, np.nan
-            r_val = np.nan
+            if cmd_type == "T":
+                self.execute_t_cmd(cmd)
+
+            elif cmd_type == "G":
+                self.excecute_g_cmd(cmd)
+
+    def excecute_m_cmd(self, cmd):
+        self.m_commands[cmd[0][1]][0]()
+
+    def execute_t_cmd(self, cmd):
+        self.t_commands[cmd[0][1]][0]()
+        self.calibrate_tool()
+
+    def excecute_g_cmd(self, cmd):
+
+        self.last_point = np.array(self.new_point)
+        self.last_or = np.array(self.new_or)
+
+        g_cmd_type = cmd[0][1]
+
+        # Activation of the zero offset
+        if g_cmd_type == 54:
+            self.offset = np.array([self.last_point, self.last_or])
+
+        # Deactivation of the zero offset
+        elif g_cmd_type == 500:
+            self.offset = np.array([[0.0, 0.0, 0.0],
+                                    [0.0, 0.0, 0.0]])
+
[…]
