### 2023-04-19 03abc0a
Renaming plane to axis
- The class variable self.plane was renamed to self.axis
