### 2023-04-19 b3104fb
Update of read_gcode
- for-loop does not need the break command
