### 2023-04-19 4567a24
Update init-Method
- else conditions was doubbled with the constructur of the method
