### 2023-04-20 a0a3a31
Update __init__
- Default values and None values updated
- test class updated according to changes
