### 2023-04-21 5a2bf30
Offset_def inside the __init__
- moved the default value of the offset inside the init method
