### 2023-04-22 #5174 @adi3
@liquidcronos please fix the Invalid DOIs as shown above
