+++ paper.bib
@@ -4,6 +4,7 @@
   booktitle={2004 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)(IEEE Cat. No. 04CH37566)},
   volume={3},
   pages={2149--2154},
+  doi = {10.1109/IROS.2004.1389727}
   organization={IEEE}
 }
 
@@ -25,7 +26,7 @@ pages = {213-229},
 year = {2014},
 note = {8th International Conference on Digital Enterprise Technology - DET 2014 Disruptive Innovation in Manufacturing Engineering towards the 4th Industrial Revolution},
 issn = {2212-8271},
-doi = {https://doi.org/10.1016/j.procir.2014.10.032},
+doi = {10.1016/j.procir.2014.10.032},
 url = {https://www.sciencedirect.com/science/article/pii/S2212827114010634},
 author = {D. Mourtzis and M. Doukas and D. Bernidaki},
 keywords = {Manufacturing, Simulation, Information and Communication Technologies},
@@ -50,7 +51,7 @@ number = {2},
 pages = {109-144},
 year = {1982},
 issn = {0146-664X},
-doi = {https://doi.org/10.1016/0146-664X(82)90169-1},
+doi = {10.1016/0146-664X(82)90169-1},
 url = {https://www.sciencedirect.com/science/article/pii/0146664X82901691},
 author = {Scott D Roth},
 abstract = {This paper presents ray casting as the methodological basis for a CAD/CAM solid modeling system. Solid objects are modeled by combining primitive solids, such as blocks and cylinders, using the set operators union, intersection, and difference. To visualize and analyze the composite solids modeled, virtual light rays are cast as probes. By virtue of its simplicity, ray casting is reliable and extensible. The most difficult mathematical problem is finding line-surface intersection points. So surfaces such as planes, quadrics, tori, and probably even parametric surface patches may bound the primitive solids. The adequacy and efficiency of ray casting are issues addressed here. A fast picture generation capability for interactive modeling is the biggest challenge. New methods are presented, accompanied by sample pictures and CPU times, to meet the challenge.}
