### 2023-05-02 22c9f7d
Changing GCodeProcessor to iterable object
Changes:
- first change of the class to be an iterable
- all the methods contain return value
- outsourcing of the read_gcode methode

Bugs:
- G-Code can be played but needs correction to fullfill the desired function
