+++ examples/g_code_processor.py
@@ -1,0 +1,289 @@
+import pybullet as p
+from pybullet_industrial import RobotBase
+from pybullet_industrial import ToolPath
+from pybullet_industrial import linear_interpolation
+from pybullet_industrial import circular_interpolation
+import numpy as np
+
+
+class GCodeProcessor:
+
+    def __init__(self, gcode=None, robot: RobotBase = None,
+                 endeffector_list: list = None,
+                 m_commands: list = None,
+                 t_commands: list = None,
+                 offset: np.array = np.array([[0.0, 0.0, 0.0],
+                                              [0.0, 0.0, 0.0]]),
+                 axis: int = 2, interpolation_steps: int = 10,
+                 sleep: int = 0.0001):
+        """Initialize a PathMover object with the provided parameters.
+
+        Args:
+            robot (pi.RobotBase, optional): The robot which is controlled.
+            filename (str, optional): The name of a G-code file to read.
+            endeffector_list (list, optional): List of end effectors to use.
+            m_commands (list, optional): M-commands to execute.
+            t_commands (list, optional): T-commands to execute.
+            offset (np.array, optional): The offset for the path points.
+            axis (int, optional): The axis around which the circle
+                                  is interpolated.Defaults to 2 which
+                                  corresponds to the z-axis (0=x,1=y)
+            interpolation_steps (int, optional): Number of interpolation steps.
+
+        Returns:
+            None
+        """
+        # Implemented for the Iteration
+        self.gcode = gcode
+        self.index = 0
+
+        self.new_point = []
+        self.new_or = []
+        self.last_point = []
+        self.last_or = []
+        self.active_endeffector = -1
+
+        self.robot = robot
+        self.m_commands = m_commands
+        self.t_commands = t_commands
+        self.endeffector_list = endeffector_list
+        self.offset = offset
+        self.axis = axis
+        self.interpolation_steps = interpolation_steps
+        self.sleep = sleep
+
+        if robot is not None:
+            self.calibrate_tool()
+
+        if endeffector_list is not None:
+            self.active_endeffector = self.get_active_endeffector()
+
+    def __iter__(self):
+        return self
+
+    def __next__(self):
+        """Runs G-Code row by row. It is necessary to create the G-Code with
+        the same structure as the readGcode(textfile) Method.
+
+        Args:
+            gcode (list): G-Code created from read_Gcode()
+
+        Returns:
+            None
+        """
+        if self.index < len(self.gcode):
+            cmd_type = self.gcode[self.index][0][0]
+
+            if cmd_type == "M":
+                cmd = self.excecute_m_cmd(self.gcode[self.index])
+
+            if cmd_type == "T":
+                cmd = self.execute_t_cmd(self.gcode[self.index])
+
+            elif cmd_type == "G":
+                cmd = self.excecute_g_cmd(self.gcode[self.index])
+            self.index += 1
+            return cmd
+        else:
+            raise StopIteration
+
+    def excecute_m_cmd(self, cmd):
+        self.m_commands[cmd[0][1]][0]()
+
+    def execute_t_cmd(self, cmd):
+        self.t_commands[cmd[0][1]][0]()
+        self.calibrate_tool()
+
+    def excecute_g_cmd(self, cmd):
+
+        self.last_point = np.array(self.new_point)
+        self.last_or = np.array(self.new_or)
+
+        g_cmd_type = cmd[0][1]
+
+        # Activation of the zero offset
+        if g_cmd_type == 54:
+            self.offset = np.array([self.last_point, self.last_or])
+
+        # Deactivation of the zero offset
+        elif g_cmd_type == 500:
+            self.offset = np.array([[0.0, 0.0, 0.0],
[…]
