### 2023-05-05 #5174 @kyleniemeyer
Hi @liquidcronos, I think the error in the BibTeX comes from a missing comma after the `doi` field in the `gazebo` entry.
