### 2023-05-05 2f384a1
fixed missing comma in gazebo entry
