+++ paper.bib
@@ -4,7 +4,7 @@
   booktitle={2004 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)(IEEE Cat. No. 04CH37566)},
   volume={3},
   pages={2149--2154},
-  doi = {10.1109/IROS.2004.1389727}
+  doi = {10.1109/IROS.2004.1389727},
   organization={IEEE}
 }
 
