### 2023-05-05 #5174 @liquidcronos
Hey @kyleniemeyer , thanks for letting me know. It should be fixed now.
The editorialbot didnt spot this error last time, is there some way i can call the bot directly in the future to avoid you and adi3 always having to check?
