### 2023-05-05 700deae
Update GCodeProcessor as Iterable
Changes:
- Draft of a new approach to treat each elementary opereation
- If a toolpath was created the iterator runs the toolpath before moving to the next command

Notations:
- This version is not debugged
