### 2023-05-05 #5174 @gkthiruvathukal
@liquidcronos Can you confirm that the institution name is rendering properly?

wbk Institute of Production Science

Specifically, is the "wbk" supposed to be lower case?
