### 2023-05-10 a72dd9a
G1 commands inserted into GCode
Changes:

(1)
- The toolpath of the G1 commands are inserted into the gcode
- afterwards they are called like G0 operations
- this enables the positions to be simulated after each step

(2)
- each G-Code operations returns a True or False value if a a step_Simulation is needed
