### 2023-05-16 9dae74d
Interpolation with Lamda Calls
This is the first adjustment of using Lambda calls for elementary operations.
In the next commits the same principle will be implemented for M-,T- and G0 function.
