### 2023-05-17 7d1a91f
Converting into elemenary operations
Changes:
- T-, M-,G0/G1/G2/G3- as elementary operations

Next Step:
- every operation elementary --> G500 etc
- better code structure
