### 2023-05-17 ff0eec5
Converting every operation into a elementary operation
- this is a draft of the new code design
- every operation in the processor is modular

Next steps:
- returning G-code line
