+++ examples/g_code_processor.py
@@ -43,13 +43,22 @@ class GCodeProcessor:
         self.active_endeffector = -1
 
         self.robot = robot
-        self.m_commands = m_commands
-        self.t_commands = t_commands
+
         self.endeffector_list = endeffector_list
         self.offset = offset
         self.axis = axis
         self.interpolation_steps = interpolation_steps
         self.sleep = sleep
+
+        self.m_commands = m_commands
+        self.t_commands = t_commands
+
+        self.g_commands = [[] for _ in range(1000)]
+        self.g_commands[54].append(lambda: self.g_54())
+        self.g_commands[500].append(lambda: self.g_500())
+        self.g_commands[17].append(lambda: self.g_17())
+        self.g_commands[17].append(lambda: self.g_18())
+        self.g_commands[17].append(lambda: self.g_19())
 
         if robot is not None:
             self.calibrate_tool()
@@ -61,6 +70,7 @@ class GCodeProcessor:
         self.operation = []
         self.index_operation = 0
         self.index_gcode = 0
+        self.path = []
         return self
 
     def __next__(self):
@@ -85,120 +95,101 @@ class GCodeProcessor:
 
             i = self.index_gcode
             cmd_type = self.gcode[i][0][0]
-            self.index_gcode += 1
+            cmd_int = self.gcode[i][0][1]
+
+            self.last_point = np.array(self.new_point)
+            self.last_or = np.array(self.new_or)
 
             if cmd_type == "M":
-                m_int = self.gcode[i][0][1]
                 self.create_elementary_actions(
-                    m_commands=self.m_commands[m_int])
+                    m_commands=self.m_commands[cmd_int])
 
             if cmd_type == "T":
-                t_int = self.gcode[i][0][1]
                 self.create_elementary_actions(
-                    t_commands=self.t_commands[t_int])
+                    t_commands=self.t_commands[cmd_int])
 
-            elif cmd_type == "G":
-                self.excecute_g_cmd(self.gcode[i])
+            elif cmd_type == "G" and cmd_int > 3:
+
+                self.create_elementary_actions(
+                    g_commands=cmd_int)
+
+            self.index_gcode += 1
 
         else:
             raise StopIteration
 
-    def create_elementary_actions(self, path: ToolPath = None, m_commands=None, t_commands=None):
+    def create_elementary_actions(self, g_int=None, m_int=None, t_int=None):
 
-        if path is not None:
+        if g_int is not None:
+            if g_int > 3:
+                for actions in self.g_commands[g_int]:
+                    self.operation.append(lambda: actions())
+            else:
+                path = self.create_path(self)
+                self.set_path(self, path)
 
-            active = self.active_endeffector  # abbreviation
-
-            for position, orientation, _ in path:
-                if self.active_endeffector == -1:
-                    self.operation.append(
-                        lambda i=position, j=orientation: self.robot.set_endeffector_pose(i, j))
-
-                else:
-                    self.operation.append(
-                        lambda i=position, j=orientation: self.endeffector_list[active].set_tool_pose(i, j))
-
-        if m_commands is not None:
-            for actions in m_commands:
+        elif m_int is not None:
+            for actions in self.m_commands[m_int]:
                 self.operation.append(lambda: actions())
 
-        if t_commands is not None:
-            for actions in t_commands:
+        elif t_int is not None:
+            for actions in self.t_commands[t_int]:
                 self.operation.append(lambda: actions())
 
             self.operation.append(lambda: self.calibrate_tool())
 
-    def excecute_g_cmd(self, cmd):
-
-        self.last_point = np.array(self.new_point)
-        self.last_or = np.array(self.new_or)
-
[…]
